Move reader package to top and use
