+++ yarn/handler.go
@@ -15,8 +15,8 @@ import (
 
 	"github.com/opensbom-generator/parsers/meta"
 	"github.com/opensbom-generator/parsers/plugin"
+	"github.com/opensbom-generator/parsers/reader"
 	"github.com/spdx/spdx-sbom-generator/pkg/helper"
-	"github.com/spdx/spdx-sbom-generator/pkg/reader"
 )
 
 type yarn struct {
